requirements.txt
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
pymongo==3.10.1
requests==2.22.0
line-bot-sdk==1.16.0